+++ src/Workspace.cpp
@@ -119,9 +119,8 @@ bool WorkspaceFolder::updateSourceMap()
 
         // Recreate instance types
         instanceTypes.clear();
-        types::createInstanceTypes(frontend.typeChecker, instanceTypes, fileResolver);
-        types::registerInstanceTypes(frontend.typeChecker, fileResolver, /* TODO - expressiveTypes: */ false);
-        types::registerInstanceTypes(frontend.typeCheckerForAutocomplete, fileResolver, /* TODO - expressiveTypes: */ true);
+        types::registerInstanceTypes(frontend.typeChecker, instanceTypes, fileResolver, /* TODO - expressiveTypes: */ false);
+        types::registerInstanceTypes(frontend.typeCheckerForAutocomplete, instanceTypes, fileResolver, /* TODO - expressiveTypes: */ true);
 
         // Signal diagnostics refresh
         client->terminateWorkspaceDiagnostics();